==> src/sprint_3/05_allresidents_admissions_discharges_regional.R <==
####
# Hospital admissions / discharges of care home residents (with MPI flag or based on ADMISORC/DISDEST)
# Process regional and care home type splits generated by RW
####

library(tidyverse)
library(tidylog)
library(janitor)
library(lubridate)
library(readxl)

library(zoo)
library(data.table)

source("file_paths.R")
source("functions.R")

# Import data -------------------------------------------------------------
# note: in contrast to the analysis in script 03, these counts of admissions and discharges include permanent residents
# as well as anyone coming from/going to care homes according to SUS
# the input files were generated by RW using the IAU SAS pipeline

discharges_region <- read_xlsx(str_c(raw_data_path, "April 2020/01b By Region exc Deaths AdmRgn Corrected.xlsx"),
                       sheet = "Disch 1420 All Exclusions")

admissions_region <- read_xlsx(str_c(raw_data_path, "April 2020/01b By Region exc Deaths AdmRgn Corrected.xlsx"),
                               sheet = "Adm 1420 All Exclusions")

# Clean & derive ----------------------------------------------------------

## Discharges
discharges_region <- discharges_region %>% 
  mutate(ch_nursing_disch = na_if(ch_nursing_disch, "NA")) %>% 
  pivot_longer(c(-dischmnth, - dischday, -rgn19nm, -ch_nursing_disch), names_to = "type", values_to = "value") %>% 
  separate(type, into = c("dischtype", "year"), sep = "_") %>% 
  mutate(date = ymd(str_c(year, dischmnth, dischday, sep = "-")), 
         day_dummy = `year<-`(date, 0004),
         destination = case_when(dischtype == "dischall" ~ "all",
                                 dischtype == "dischother" ~ "other",
                                 dischtype == "dischch" & ch_nursing_disch == 0 ~ "res_care_home",
                                 dischtype == "dischch" & ch_nursing_disch == 1 ~ "nursing_home",
                                 dischtype == "dischch" & is.na(ch_nursing_disch) ~ "unknown_care_home",))

# remove some variable combinations that aren't defined
discharges_region <- discharges_region %>% 
  filter(!(destination %in% c("all", "other") & !is.na(ch_nursing_disch)))

discharges_region <- discharges_region %>% 
  select(-ch_nursing_disch, -dischtype, -dischmnth, -dischday) %>% 
  pivot_wider(names_from = "destination", values_from = "value") 

discharges_region <- discharges_region %>% 
  mutate(all_recalc = rowSums(select(., one_of("res_care_home", "nursing_home", "other", "unknown_care_home")), na.rm = TRUE),
         all_match = ifelse(all == all_recalc, 1, 0),
         all_care_home = rowSums(select(., one_of("res_care_home", "nursing_home", "unknown_care_home")), na.rm = TRUE))

## Admissions
admissions_region <- admissions_region %>% 
  mutate(ch_nursing_adm = na_if(ch_nursing_adm, "NA")) %>% 
  pivot_longer(c(-admmnth, - admday, -rgn19nm, -ch_nursing_adm), names_to = "type", values_to = "value") %>% 
  separate(type, into = c("admtype", "year"), sep = "_") %>% 
  mutate(date = ymd(str_c(year, admmnth, admday, sep = "-")), 
         day_dummy = `year<-`(date, 0004),
         source = case_when(admtype == "admall" ~ "all",
                            admtype == "admother" ~ "other",
                            admtype == "admch" & ch_nursing_adm == 0 ~ "res_care_home",
                            admtype == "admch" & ch_nursing_adm == 1 ~ "nursing_home",
                            admtype == "admch" & is.na(ch_nursing_adm) ~ "unknown_care_home",))

# remove some variable combinations that aren't defined
admissions_region <- admissions_region %>% 
  filter(!(source %in% c("all", "other") & !is.na(ch_nursing_adm)))

admissions_region <- admissions_region %>% 
  select(-ch_nursing_adm, -admtype, -admmnth, -admday) %>% 
  pivot_wider(names_from = "source", values_from = "value") 

admissions_region <- admissions_region %>% 
  mutate(all_recalc = rowSums(select(., one_of("res_care_home", "nursing_home", "other", "unknown_care_home")), na.rm = TRUE),
         all_match = ifelse(all == all_recalc, 1, 0),
         all_care_home = rowSums(select(., one_of("res_care_home", "nursing_home", "unknown_care_home")), na.rm = TRUE))




# Save aggregates ---------------------------------------------------------

## Discharges
# regional care home discharges during COVID period 2019-2020
discharges_ch_COVIDperiod <- discharges_region %>% 
  filter(day_dummy > ymd("0004-03-17") &  day_dummy < ymd("0004-04-30") &year %in% c(2019, 2020)) %>% 
  group_by(year, rgn19nm) %>% 
  summarise(disch_res_care_home = sum(res_care_home, na.rm = TRUE),
            disch_nursing_home = sum(nursing_home, na.rm = TRUE),
            disch_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            disch_all_care_homes = sum(all_care_home, na.rm = TRUE)) %>% 
  pivot_longer(c(-year, - rgn19nm), names_to = "dischtype", values_to = "count") %>% 
  group_by(rgn19nm, dischtype) %>% 
  mutate(pct_change_2019 = round(100 * count / count[year == "2019"], 1))

write_csv(discharges_ch_COVIDperiod, str_c(results_path_sprint3, "CH_discharges_regional_Mar-April.csv"))

# national care home discharges during March/April period 2015-2020
discharges_ch_COVIDperiod_national <- discharges_region %>% 
  filter(day_dummy >= ymd("0004-03-01") & day_dummy <= ymd("0004-04-30")) %>% 
  group_by(year) %>% 
  summarise(disch_res_care_home = sum(res_care_home, na.rm = TRUE),
            disch_nursing_home = sum(nursing_home, na.rm = TRUE),
            disch_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            disch_all_care_homes = sum(all_care_home, na.rm = TRUE),
            date_start = min(date),
            date_end = max(date)) %>% 
  pivot_longer(c(-year, - date_start, -date_end), names_to = "dischtype", values_to = "count") %>% 
  group_by(dischtype) %>% 
  mutate(pct_change_2019 = round(100 * count / count[year == "2019"], 1),
         ref_mean_2015_to_2019 = mean(count[year %in% c("2015", "2016", "2017", "2018", "2019")]),
         pct_change_2015_to_2019 = round(100 * count / ref_mean_2015_to_2019, 1))

write_csv(discharges_ch_COVIDperiod_national, str_c(results_path_sprint3, "CH_discharges_national_Mar-April.csv"))

# national care home discharges during COVID period 2015-2020
 discharges_region %>% 
  filter(day_dummy >= ymd("0004-03-17") & day_dummy <= ymd("0004-04-30")) %>% 
  group_by(year) %>% 
  summarise(disch_res_care_home = sum(res_care_home, na.rm = TRUE),
            disch_nursing_home = sum(nursing_home, na.rm = TRUE),
            disch_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            disch_all_care_homes = sum(all_care_home, na.rm = TRUE),
            date_start = min(date),
            date_end = max(date)) %>% 
  pivot_longer(c(-year, - date_start, -date_end), names_to = "dischtype", values_to = "count") %>% 
  group_by(dischtype) %>% 
  mutate(pct_change_2019 = round(100 * count / count[year == "2019"], 1),
         ref_mean_2015_to_2019 = mean(count[year %in% c("2015", "2016", "2017", "2018", "2019")]),
         pct_change_2015_to_2019 = round(100 * count / ref_mean_2015_to_2019, 1)) %>% 
   write_csv(., str_c(results_path_sprint3, "CH_discharges_national_Mar17-April.csv"))
 

# Visualise regional time trends
discharges_region %>% 
  arrange(date) %>% 
  group_by(rgn19nm) %>% 
  mutate(all_care_home_sevendayavg = sevendayavg(all_care_home)) %>%  
  ggplot(aes(x = day_dummy, y = all_care_home_sevendayavg, group = rgn19nm, color = rgn19nm)) +
  geom_line() +
  facet_wrap("year", ncol = 2)

# National aggregates by day  2019-2020
discharges_national <- discharges_region %>% 
  filter(year %in% c(2019, 2020)) %>% 
   group_by(year, date, day_dummy) %>% 
   summarise(all = sum(all_recalc),
             all_care_home = sum(all_care_home),
             all_other = sum(other))

## Admissions
# regional care home admissions during COVID period 2019-2020
admissions_ch_COVIDperiod <- admissions_region %>% 
  filter(day_dummy > ymd("0004-03-17")&  day_dummy < ymd("0004-04-30")  & year %in% c(2019, 2020)) %>% 
  group_by(year, rgn19nm) %>% 
  summarise(adm_res_care_home = sum(res_care_home, na.rm = TRUE),
            adm_nursing_home = sum(nursing_home, na.rm = TRUE),
            adm_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            adm_all_care_homes = sum(all_care_home, na.rm = TRUE)) %>% 
  pivot_longer(c(-year, - rgn19nm), names_to = "adm_type", values_to = "count") %>% 
  group_by(rgn19nm, adm_type) %>% 
  mutate(pct_change_2019 = round(100 * count / count[year == "2019"], 1))

write_csv(admissions_ch_COVIDperiod, str_c(results_path_sprint3, "CH_admissions_regional_Mar-April.csv"))

# national care home admissions during March/April period 2015-2020
admissions_ch_COVIDperiod_national <- admissions_region %>% 
  filter(day_dummy >= ymd("0004-03-01") & day_dummy <= ymd("0004-04-30")) %>% 
  group_by(year) %>% 
  summarise(adm_res_care_home = sum(res_care_home, na.rm = TRUE),
            adm_nursing_home = sum(nursing_home, na.rm = TRUE),
            adm_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            adm__all_care_homes = sum(all_care_home, na.rm = TRUE),
            date_start = min(date),
            date_end = max(date)) %>% 
  pivot_longer(c(-year, - date_start, -date_end), names_to = "adm_type", values_to = "count") %>% 
  group_by(adm_type) %>% 
  mutate(pct_change_2019 = round(100 * count / count[year == "2019"], 1),
         ref_mean_2015_to_2019 = mean(count[year %in% c("2015", "2016", "2017", "2018", "2019")]),
         pct_change_2015_to_2019 = round(100 * count / ref_mean_2015_to_2019, 1))

write_csv(admissions_ch_COVIDperiod_national, str_c(results_path_sprint3, "CH_admissions_national_Mar-April.csv"))

# national care home admissions during COVID period 2015-2020
admissions_region %>% 
  filter(day_dummy >= ymd("0004-03-17") & day_dummy <= ymd("0004-04-30")) %>% 
  group_by(year) %>% 
  summarise(adm_res_care_home = sum(res_care_home, na.rm = TRUE),
            adm_nursing_home = sum(nursing_home, na.rm = TRUE),
            adm_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            adm__all_care_homes = sum(all_care_home, na.rm = TRUE),
            date_start = min(date),
            date_end = max(date)) %>% 
  pivot_longer(c(-year, - date_start, -date_end), names_to = "adm_type", values_to = "count") %>% 
  group_by(adm_type) %>% 
  mutate(pct_change_2019 = round(100 * count / count[year == "2019"], 1),
         ref_mean_2015_to_2019 = mean(count[year %in% c("2015", "2016", "2017", "2018", "2019")]),
         pct_change_2015_to_2019 = round(100 * count / ref_mean_2015_to_2019, 1)) %>% 
  write_csv(., str_c(results_path_sprint3, "CH_admissions_national_Mar17-April.csv"))

# Visualise regional time trends 
admissions_region %>% 
  arrange(date) %>% 
  group_by(rgn19nm) %>% 
  mutate(all_care_home_sevendayavg = sevendayavg(all_care_home)) %>%  
  ggplot(aes(x = day_dummy, y = all_care_home_sevendayavg, group = rgn19nm, color = rgn19nm)) +
  geom_line() +
  facet_wrap("year", ncol = 2)

# National aggregates by day 2019-2020
admissions_national <- admissions_region %>% 
  filter(year %in% c(2019, 2020)) %>% 
  group_by(year, date, day_dummy) %>% 
  summarise(all = sum(all_recalc),
            all_care_home = sum(all_care_home),
            all_other = sum(other))



# Weekly aggregates for care home modelling -------------------------------


admissions_region %>% 
  filter(year %in% c("2016", "2020")) %>% 
  mutate(week = isoweek(date))  %>% 
  group_by(year, week) %>% 
  summarise(adm_res_care_home = sum(res_care_home, na.rm = TRUE),
            adm_nursing_home = sum(nursing_home, na.rm = TRUE),
            adm_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            adm_all_care_homes = sum(all_care_home, na.rm = TRUE),
            date_start = min(date),
            date_end = max(date)) %>% 
  write_csv(., str_c(results_path_sprint3, "CH_admissions_national_weekly_2016_2020.csv"))

admissions_region %>% 
  filter(date %within% interval(ymd("2016-03-28"), ymd("2017-04-02"))) %>% 
  mutate(week = isoweek(date))  %>% 
  filter(week != 52) %>% 
  group_by(year, week) %>% 
  summarise(adm_res_care_home = sum(res_care_home, na.rm = TRUE),
            adm_nursing_home = sum(nursing_home, na.rm = TRUE),
            adm_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            adm_all_care_homes = sum(all_care_home, na.rm = TRUE),
            date_start = min(date),
            date_end = max(date)) %>% 
  ungroup() %>% 
  summarise(n_weeks = n(),
            date_min = min(date_start),
            date_max = max(date_start),
            adm_res_care_home = mean(adm_res_care_home, na.rm = TRUE),
            adm_nursing_home = mean(adm_nursing_home, na.rm = TRUE),
            adm_unknown_care_home = mean(adm_unknown_care_home, na.rm = TRUE),
            adm_all_care_homes = mean(adm_all_care_homes, na.rm = TRUE)) %>% 
  write_csv(., str_c(results_path_sprint3, "CH_admissions_national_avg_weekly_2016-2017.csv"))
  
discharges_region %>% 
  filter(year %in% c("2016", "2020")) %>% 
  mutate(week = isoweek(date))  %>% 
  group_by(year, week) %>% 
  summarise(disch_res_care_home = sum(res_care_home, na.rm = TRUE),
            disch_nursing_home = sum(nursing_home, na.rm = TRUE),
            disch_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            disch_all_care_homes = sum(all_care_home, na.rm = TRUE),
            date_start = min(date),
            date_end = max(date)) %>% 
  write_csv(., str_c(results_path_sprint3, "CH_discharges_national_weekly_2016_2020.csv"))

discharges_region %>% 
  filter(date %within% interval(ymd("2016-03-28"), ymd("2017-04-02"))) %>% 
  mutate(week = isoweek(date))  %>% 
  filter(week != 52) %>% 
  group_by(year, week) %>% 
  summarise(disch_res_care_home = sum(res_care_home, na.rm = TRUE),
            disch_nursing_home = sum(nursing_home, na.rm = TRUE),
            disch_unknown_care_home = sum(unknown_care_home, na.rm = TRUE),
            disch_all_care_homes = sum(all_care_home, na.rm = TRUE),
            date_start = min(date),
            date_end = max(date)) %>% 
  ungroup() %>% 
  summarise(n_weeks = n(),
            date_min = min(date_start),
            date_max = max(date_start),
            disch_res_care_home = mean(disch_res_care_home, na.rm = TRUE),
            adisch_nursing_home = mean(disch_nursing_home, na.rm = TRUE),
            disch_unknown_care_home = mean(disch_unknown_care_home, na.rm = TRUE),
            disch_all_care_homes = mean(disch_all_care_homes, na.rm = TRUE)) %>% 
  write_csv(., str_c(results_path_sprint3, "CH_discharges_national_avg_weekly_2016-2017.csv"))
